# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ToDo App",
        "description": "A ToDo server writen in NodeJS with Express & Mongoose",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "tags": [
        {
            "name": "Tasks",
            "description": "Manipulating Tasks"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/x-www-form-urlencoded"
    ],
    "produces": [
        "application/json",
        "application/x-www-form-urlencoded"
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags":[
                    "Tasks"
                ],
                "summary": "Get Active Tasks",
                "description": "Return all active Tasks",
                "responses": {
                    "200": {
                        "description": "Returned successfully",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "400": {
                        "description": "Internal Error"
                    }
                }
            },
            "post": {
                "tags":[
                    "Tasks"
                ],
                "summary": "Add Task",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Task that we want to create",
                        "example": "title=title&description=description",
                        "Schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200":{
                        "description": "New Task is created",
                        "schema":{
                            "$ref":"#/definitions/Task"
                        }
                    },
                    "400":{
                        "description":"Internal Error"
                    }
                }
            }
        },
        "/tasks/deleted": {
            "get": {
                "tags":["Tasks"],
                "summary":"Get Deleted Tasks",
                "responses": {
                    "200":{
                        "description": "Returned successfully",
                        "schema":{
                            "$ref":"#/definitions/Task"
                        }
                    },
                    "400":{
                        "description":"Internal Error"
                    }
                }
            }
        },
        "/tasks/{taskId}":{
            "parameters": [
                {
                  "name": "taskId",
                  "in": "path",
                  "required": true,
                  "description": "ID of task that we want to find",
                  "type": "string"
                }
            ],
            "patch": {
                "summary": "Edit Task",
                "description": "Edit Title and Description of task",
                "tags": ["Tasks"],
                "parameters": [
                    {
                    "name": "body",
                    "in": "body",
                    "description": "Task that we want to create",
                    "example": "title=title&description=description",
                    "Schema": {
                        "$ref": "#/definitions/Task"
                    }
                }
                ],
                "responses": {
                    "200":{
                        "description": "Deleted successfully",
                        "schema":{
                            "$ref":"#/definitions/Task"
                        }
                    },
                    "400":{
                        "description":"Internal Error"
                    }
                }
            },
            "delete": {
                "summary": "Delete Task",
                "description": "Soft delete task, still available via /tasks/deleted",
                "tags": ["Tasks"],
                "responses": {
                    "200":{
                        "description": "Deleted successfully",
                        "schema":{
                            "$ref":"#/definitions/Task"
                        }
                    },
                    "400":{
                        "description":"Internal Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Task": {
            "required": [
                "title"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "unique": true
                },
                "title": {
                    "type":"string",
                    "required": true
                },
                "description": {
                    "type":"string"
                },
                "created_at": {
                    "type":"string",
                    "default": "Date.now()"
                },
                "updated_at": {
                    "type":"string",
                    "default": "Date.now()"
                },
                "deleted_at": {
                    "type":"string",
                    "default": "undefined"
                }
            }
        }
    }
}